working with authorization

+++ src/tasks/tasks.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/auth/auth/user.entity';
 import { v4 as uuidv4 } from 'uuid';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
@@ -19,6 +20,12 @@ export class TasksService {
     getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
         return this.taskRepository.getTasks(filterDto)
     }
+
+
+    getTasksByUser(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+        return this.taskRepository.getTasksByUser(filterDto, user)
+    }
+    
     
     
 
@@ -32,24 +39,26 @@ export class TasksService {
         return found;
     }
 
-    // getTaskById(id: string): Task {
-    //     const found = this.tasks.find((task: Task) => task.id === id);
-    //     if (!found) {
-    //        throw new NotFoundException(`Task with ID: ${id} not found !`);     
-    //     }
-    //     return found;
-    // }
+    async getTaskByIdAndUser(id: number, user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({ where: { id, userId: user.id} });
 
+        if (!found) {
+            throw new NotFoundException(`Task with ID: ${id} not found !`);
+        }
 
-    createTask(createTaskDto: CreateTaskDto) {
-        return this.taskRepository.createTask(createTaskDto);
+        return found;
     }
 
 
-    async deleteTask(id: number): Promise<void> {
+    createTask(createTaskDto: CreateTaskDto, user: User) {
+        return this.taskRepository.createTask(createTaskDto, user);
+    }
+
+
+    async deleteTask(id: number, user: User): Promise<void> {
         //const found = await this.getTaskById(id);
        
-        const result = await this.taskRepository.delete(id);
+        const result = await this.taskRepository.delete({id, userId: user.id});
 
         if (result.affected === 0) {
             throw new NotFoundException(`Task with ID: ${id} not found !`);
@@ -59,8 +68,8 @@ export class TasksService {
         
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const taskToUpdate = await this.getTaskById(id);
+    async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const taskToUpdate = await this.getTaskByIdAndUser(id, user);
 
         taskToUpdate.status = status;
         await taskToUpdate.save();
